Use TryFrom pattern
