+++ src/lib.rs
@@ -73,7 +73,7 @@ pub fn run_with_decoration(text: &str, config: &config::Configuration, debug: bo
 }
 
 fn run_command(text: &str, configuration: &config::Configuration, last_command: &mut Option<String>) -> Result<bool> {
-  let trigger = Trigger::parse(text)?;
+  let trigger = Trigger::try_from(text)?;
   let command = match configuration.get_command(&trigger, last_command) {
     Err(err) => match err {
       UserError::NoCommandToRepeat => {
